+++ packages/start-aws/entry.mjs
@@ -1,3 +1,4 @@
+import { splitCookiesString } from "solid-start/node/fetch";
 import "solid-start/node/globals.js";
 import manifest from "../../dist/client/route-manifest.json";
 import server from "./entry-server";
@@ -14,6 +15,11 @@ export async function handler(event) {
   for (const [name, value] of response.headers) {
     headers[name] = value;
   }
+  if (webRes.headers.has('set-cookie')) {
+		const header = /** @type {string} */ (webRes.headers.get('set-cookie'));
+		// @ts-expect-error
+		headers['set-cookie'] =  splitCookiesString(header);
+	}
 
   return {
     statusCode: response.status,
